refactor: 폴더 페이지 컴포넌트에서 폴더 목록을 호출하도록 리팩토링
